+++ src/modules/Dashboard/SignInButton.tsx
@@ -1,42 +1,12 @@
 import { gql, useMutation } from '@apollo/client';
 import { useLogin, useLogout, usePrivy } from '@privy-io/react-auth';
 import { Button } from '@site/src/components/Button';
+import { useSignIn } from '@site/src/helpers/useSignIn';
 import DropdownNavbarItem from '@theme/NavbarItem/DropdownNavbarItem';
 
-const UPSERT_USER = gql`
-  mutation UpsertApiClient($email: String!) {
-    upsertApiClient(email: $email) {
-      id
-      name
-    }
-  }
-`;
-
 export function SignInButton() {
-  const [upsertUser] = useMutation(UPSERT_USER);
   const { authenticated, ready, user } = usePrivy();
-  const { logout } = useLogout();
-
-  const { login } = useLogin({
-    onComplete: async (user, isNewUser) => {
-      if (user.email) {
-        const { data } = await upsertUser({
-          variables: {
-            email: user.email.address,
-          },
-        });
-
-        const clientId = data?.upsertApiClient?.id || '';
-        if (clientId) {
-          localStorage.setItem('zapper.clientId', clientId);
-        }
-      }
-
-      if (isNewUser) {
-        window.location.reload();
-      }
-    },
-  });
+  const { login, logout } = useSignIn();
 
   if (!ready) {
     return null;
